creates itemslist separately and reference in new stock schema
