+++ src/inventory-items-list/services/index.ts
@@ -1,0 +1,1 @@
+export * from "./items-list.service"
