+++ src/inventory-new-stock/DTOs/NewStockDTO.ts
@@ -1,16 +1,7 @@
 import { ObjectId } from "mongodb";
-import { Units } from "../../types/user";
 
 
 export type AddNewStockDTO = {
     userID: ObjectId;
-    itemList: {
-        productID: ObjectId;
-        categoryID: ObjectId;
-        brandID: ObjectId;
-        quantity: number;
-        price: number;
-        unit: Units;
-        isInStock: boolean 
-    }[]
+    itemsListID: ObjectId;
 }
