+++ src/inventory-new-stock/api/routes/index.ts
@@ -1,4 +1,4 @@
-import ProductValidation from "../validations";
+import AddNewStockValidation from "../validations";
 import { baseRouter, baseValidation } from "../../../inventory-shared/api";
 
 import { NewStockController } from "../controllers/new-stock";
@@ -6,6 +6,6 @@ import { AuthMiddleware } from "../../../inventory-auth/middlewares/authMiddlewa
 
 const {POST, router} = baseRouter();
 
-POST("/", [baseValidation(ProductValidation.newStock), AuthMiddleware.baseAuthToken, AuthMiddleware.IsUserMiddleware, NewStockController.addNewStock]);
+POST("/", [baseValidation(AddNewStockValidation.newStock), AuthMiddleware.baseAuthToken, AuthMiddleware.IsUserMiddleware, NewStockController.addNewStock]);
 
 export default router;
