+++ src/inventory-subcategory/api/routes/index.ts
@@ -1,7 +1,7 @@
 import CategoryValidation from "../validations";
 import { baseRouter, baseValidation } from "../../../inventory-shared/api";
 
-import { SubCategoryController } from "../controllers/subcategory";
+import { SubCategoryController } from "../controllers/subCategory";
 import { AuthMiddleware } from "../../../inventory-auth/middlewares/authMiddleware";
 
 const {POST, router} = baseRouter();
