+++ src/inventory-web-api/api/routes.ts
@@ -11,7 +11,7 @@ import { Request, Response, Router } from 'express';
  import subCategoryRoute from '../../inventory-subcategory/api/routes';
  import brandRoute from '../../inventory-brand/api/routes';
  import productRoute from '../../inventory-product/api/routes';
- import newStockRoute from '../../inventory-new-stock/api/routes';
+ import itemsListRoute from '../../inventory-items-list/api/routes';
  import staffOrderRoute from '../../inventory-staff-order/api/routes';
 
 
@@ -30,7 +30,7 @@ router.use('/category', categoryRoute);
 router.use('/subcategory', subCategoryRoute);
 router.use('/brand', brandRoute);
 router.use('/product', productRoute);
-router.use('/new-stock', newStockRoute);
+router.use('/items-list', itemsListRoute);
 router.use('/staff-order', staffOrderRoute);
